+++ src/demo/BasicState/BasicState.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Provider } from "../basicTruth";
+import { Name } from "./Name";
+import { List } from "./List";
+
+const BasicStateDemo = () => {
+    return (
+        <Provider>
+            <Name />
+            <List />
+        </Provider>
+    );
+};
+
+export { BasicStateDemo };