+++ src/index.spec.js
@@ -1,5 +1,0 @@
-describe("tests", () => {
-    it("are working just fine", () => {
-        true.must.be.true();
-    });
-});