+++ src/index.tsx
@@ -1,0 +1,2 @@
+export * from "./createTruth.js";
+export * from "./truth.js";