+++ src/pages/BasicState.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { BasicStateDemo } from "../demo/BasicState/BasicState";
+
+const BasicState = () => {
+    return (
+        <BasicStateDemo />
+    );
+};
+
+export default BasicState;