+++ src/pages/Legacy.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { LegacyStateDemo } from "../demo/Legacy/Legacy";
+
+const LegacyState = () => {
+    return (
+        <LegacyStateDemo />
+    );
+};
+
+export default LegacyState;